Moved server selection to own file. Renamed certain variables
